move useMyUserGroups to GrantSelector directory
