+++ apps/app/src/components/SavePageControls/GrantSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './GrantSelector';
